% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hr_forestly.R
\name{format_hr_forestly}
\alias{format_hr_forestly}
\title{Format the outdata for the HR forest plot, along with its KM plots}
\usage{
format_hr_forestly(
  outdata,
  display = c("n", "event", "fig_hr"),
  digits = 2,
  width_subgroup = 50,
  width_fig = 360,
  width_n = 40,
  width_event = 40,
  width_hr = 40,
  footer_space = 150,
  hr_range = NULL,
  color = NULL,
  hr_label = "Treatment <- Favor -> Placebo"
)
}
\arguments{
\item{outdata}{An \code{outdata} object created by \code{\link[=prepare_hr_forestly]{prepare_hr_forestly()}}.}

\item{display}{A character vector specifying which columns to display in the forest plot.
Options include "n", "event", "hr", and "fig_hr".
Default is c("n", "event", "fig_hr").
\itemize{
\item \code{n}: Number of participants in a comparison.
\item \code{event}: Number of events in a comparison.
\item \code{hr}: Hazard ratio estimate in a comparison.
\item \code{fig_hr}: Hazard ratio figure in a comparison.
}}

\item{digits}{A numeric value specifying the number of digits to
display for hazard ratios and confidence intervals. Default is 2.}

\item{width_subgroup}{A numeric value specifying the width of
the subgroup column in pixels.}

\item{width_fig}{A numeric value specifying the width of
the hazard ratio figure column in pixels.}

\item{width_n}{A numeric value specifying the width of the "n" column in pixels.}

\item{width_event}{A numeric value specifying the width of
the "event" column in pixels.}

\item{width_hr}{A numeric value specifying the width of the hazard ratio column in pixels.}

\item{footer_space}{A numeric value specifying the space for the footer in pixels.}

\item{hr_range}{A numeric vector of lower and upper limit of x-axis
for the hazard ratio figure.}

\item{color}{A character vector of colors to use for the hazard ratio figures.
Defaulte value supports up to 4 groups.}

\item{hr_label}{A character string specifying the label for the hazard ratio axis.}
}
\value{
An \code{outdata} object.
}
\description{
Format the outdata for the HR forest plot, along with its KM plots
}
\examples{
prepare_hr_forestly(
  meta = meta_tte_example_new(),
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;os",
  subgroup = "gender",
  km_curves = "female;male",
  arm_levels = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose")
) |>
  format_hr_forestly()
}
