% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_forestly.R
\name{hr_forestly}
\alias{hr_forestly}
\title{Display the HR forest plot, along with its KM plots}
\usage{
hr_forestly(
  outdata,
  time_unit = c("days", "weeks", "months", "years"),
  y_label_kmplot = "Survival Probability",
  width = 1000,
  height_kmoplot = 400,
  height_at_risk = 200,
  max_page = NULL
)
}
\arguments{
\item{outdata}{outdata An \code{outdata} object created by \code{\link[=format_hr_forestly]{format_hr_forestly()}}.}

\item{time_unit}{A string specifying the unit of time in x-axis of the KM plot.
must be one of "days", "weeks", "months", or "years". Default is "days".}

\item{y_label_kmplot}{A string specifying the y-axis label of the KM plot.
Default is "Survival Probability".}

\item{width}{A numeric value of width of the entire plot in pixels. Default is 1000.}

\item{height_kmoplot}{A numeric value of height of the KM plot in pixels. Default is 400.}

\item{height_at_risk}{A numeric value of height of the at-risk table in pixels.
Default is 200.}

\item{max_page}{A numeric value of maximum number of subgroups to display per page.}
}
\value{
An HR forest plot with subgroup KM plot saved as a \code{shiny.tag.list} object.
}
\description{
Display the HR forest plot, along with its KM plots
}
\examples{
prepare_hr_forestly(
  meta = meta_tte_example_new(),
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;os",
  subgroup = "gender",
  km_curves = "male;female",
  arm_levels = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose")
) |>
  format_hr_forestly() |>
  hr_forestly()
}
