% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_extract.R
\name{km_extract}
\alias{km_extract}
\title{Extract Survival Estimates}
\usage{
km_extract(f.survfit)
}
\arguments{
\item{f.survfit}{Survival fit object from survfit.}
}
\value{
Data frame of timepoint estimates from km methods.
}
\description{
This function extracts the km estimates from survfit.
}
\section{Specification}{

\if{latex}{

Derivation logic in \code{km_extract}:

 \itemize{
 \item Check if f.survfit argument type is valid.
 \item Summarize km estimates.
 \item Manipulate and define strata name.
 \item Define output data frame with:
 time, timepoint from km raw estimates;
 n.risk, number at risk at the specific timepoint;
 n.event, number of events at the specific timepoint;
 n.censor, number of censored patients at the specific timepoint;
 surv, survival probability from km estimates at the time point;
 upper/lower, upper and lower limit of the km estimates (survival rate);
 strata, the strata of the population (from the survfit), if no just use the 'All';
 strata_name, name of strata variable to be displayed.
 }
 }

\if{html}{The contents of this section are shown in PDF user manual only.}
}

\keyword{internal}
