% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcurvely.R
\name{kmcurvely}
\alias{kmcurvely}
\title{Generate interactive KM plot by endpoints and subgroups}
\usage{
kmcurvely(
  meta = meta_tte_example(),
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;os",
  subgroup = "male;female",
  x_label = NULL,
  y_label = "Survival rate",
  color = NULL,
  time_unit = c("days", "weeks", "months", "years")
)
}
\arguments{
\item{meta}{A metadata object that contains the mappings for the variables of interest.}

\item{population}{A string indicating the population to use (e.g., "apat").}

\item{observation}{A string indicating the observation population (e.g., "efficacy_population").}

\item{endpoint}{A semicolon-separated string of endpoints to be analyzed (e.g., "pfs;os").}

\item{subgroup}{A semicolon-separated string of subgroups to filter by (e.g., "male;female").}

\item{x_label}{A string for the x-axis label (default is "Time in days/weeks/months/years").}

\item{y_label}{A string for the y-axis label (default is "Survival rate").}

\item{color}{A vector of colors for the plot (if NULL, defaults to a preset color palette).}

\item{time_unit}{A string specifying the unit of time; must be one of "days", "weeks", "months", or "years". Default is "days".}
}
\value{
An interactive KM plot
}
\description{
Generate interactive KM plot by endpoints and subgroups
}
\examples{
# Example usage of kmcurvely function

meta <- meta_tte_example()
kmcurvely(meta = meta,
           population = "apat",
           observation = "efficacy_population",
           endpoint = "pfs;os",
           subgroup = "male;female")


}
