% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kmcurvely_adsl}
\alias{kmcurvely_adsl}
\title{ADSL dataset}
\format{
A data frame with 254 rows and 48 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
kmcurvely_adsl
}
\value{
An analysis data frame.
}
\description{
A dataset containing subject level information.
}
\keyword{datasets}
