% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kmcurvely_adtte}
\alias{kmcurvely_adtte}
\title{ADTTE dataset}
\format{
A data frame with 254 rows and 26 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
kmcurvely_adtte
}
\value{
An analysis data frame.
}
\description{
A dataset containing Time-to-Event Analysis details.
}
\keyword{datasets}
