% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcurvely_static.R
\name{kmcurvely_static}
\alias{kmcurvely_static}
\title{Generate Static KM plot by endpoints and subgroups and save as RTF/PNG files}
\usage{
kmcurvely_static(
  surv_shared,
  tbl_at_risk,
  x_label = "Time in Weeks",
  y_label = "Survival rate",
  color,
  xlimit,
  break_x_by,
  population,
  legned_pos = c(0.1, 0.12),
  folder_path = tempdir(),
  zipname = "all_static_plots.zip"
)
}
\arguments{
\item{surv_shared}{A data frame of survival data per endpoint per subgroup}

\item{tbl_at_risk}{A data frame containing the number of subjects at risk}

\item{x_label}{The label for the x-axis}

\item{y_label}{The label for the y-axis}

\item{color}{The color palette for different treament groups}

\item{xlimit}{The maximum value for the x-axis}

\item{break_x_by}{The interval for the x-axis breaks}

\item{population}{The population specify in kmcurvely.r}

\item{legned_pos}{The position of the legend}

\item{folder_path}{The path to save the RTF files}

\item{zipname}{The name of the zip file for all static KM plots}
}
\value{
A zip file containing RTF files of the static KM plots
}
\description{
Generate Static KM plot by endpoints and subgroups and save as RTF/PNG files
}
\examples{
key_data_objects <- attr(kmcurvely(), "key_data_objects")

kmcurvely_static(
  surv_shared = key_data_objects$tbl_surv,
  tbl_at_risk = key_data_objects$tbl_at_risk,
  x_label = key_data_objects$x_label,
  y_label = key_data_objects$y_label,
  color = key_data_objects$color,
  xlimit = key_data_objects$xlimit,
  break_x_by = key_data_objects$break_x_by,
  population = key_data_objects$population
)
}
