% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_tte_example_new.R
\name{meta_tte_example_new}
\alias{meta_tte_example_new}
\title{Example Function for Time-to-Event Analysis}
\usage{
meta_tte_example_new()
}
\value{
A list containing metadata structured for time-to-event analysis.
}
\description{
This function prepares the metadata required for conducting time-to-event
analysis using Kaplan-Meier curves in the context of a clinical trial.
It cleans and processes the datasets, defines the analysis plan, and
structures the metadata necessary for further analysis.
}
\examples{
meta_data <- meta_tte_example_new()
# Use the returned meta_data in subsequent analysis functions.
}
