% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hr_forestly.R
\name{prepare_hr_forestly}
\alias{prepare_hr_forestly}
\title{Prepare the dataset for the HR forest plot, alongwith its KM plots}
\usage{
prepare_hr_forestly(
  meta = NULL,
  population = NULL,
  observation = NULL,
  endpoint = NULL,
  subgroup = NULL,
  arm_levels = NULL
)
}
\arguments{
\item{meta}{A metadata object that contains the mappings for the variables of interest.}

\item{population}{A string indicating the population to use (e.g., "apat").}

\item{observation}{A string indicating the observation population (e.g., "efficacy_population").}

\item{endpoint}{A semicolon-separated string of endpoints to be analyzed (e.g., "pfs;os").}

\item{subgroup}{A semicolon-separated string of subgroups to filter by (e.g., "male;female").}

\item{arm_levels}{A vector of character specifying the levels of arms, starting with the reference arm.}
}
\value{
An metadata with HR per subgroup along with its KM plotting data
}
\description{
Prepare the dataset for the HR forest plot, alongwith its KM plots
}
\examples{
prepare_hr_forestly(
  meta = meta_tte_example_new(),
  population = "apat",
  observation = "efficacy_population",
  endpoint = "pfs;os",
  subgroup = "male;female",
  arm_levels = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose")
)
}
